moved settings view to it own file
